Ignore certain instances
Management cluster might be in turn be managed by another cluster.

In such scenario:
1.there is an instance of sveltos-agent running in the
management cluster;
2. there are two types of EventSource/HealthCheck/Classifier instances.
One that is defined in the management cluster to be pushed and evaluated
in the managed cluster. One that is pushed by the other cluster. Such
instances need to be evaluated by the sveltos-agent running in the
management cluster.

In order to properly handle such scenario:
1. when an EventSource/HealthCheck/Classifier instance has annotation
set by sveltos, agent will process it
2. otherwise sveltos-agent will ignore it