Run sveltos-agent in the management cluster
When running in the management cluster, there will be one
sveltos-agent per cluster.
Classifier will start (and delete) those deployments accordingly.

When running in the management cluster, sveltos-agent will get the
managed cluster kubeconfig.

The `current-cluster` option is how sveltos-agent is told whether
it is running in the managed or management cluster.