Classifier
With respect to resources, Classifier can now classify a cluster
using two fields:

- ResourceSelectors
- AggregatedClassification.

If AggregatedClassification is not provided, a cluster is considered a match if and
only if each ResourceSelector finds at least one matching resource.
Otherwise, the function first collects the resources that match each ResourceSelector
and then passes all of these resources to the AggregatedClassification function for
further evaluation.

AggregatedClassification must contain a function named "evaluate"
which returns a struct with:
- "matching" field: boolean indicating whether cluster is a match;
- "message" field: (optional) message.